+++ examples/resources/artifactory_local_nix_repository/resource.tf
@@ -1,0 +1,5 @@
+resource "artifactory_local_nix_repository" "my-nix-local" {
+  key         = "my-nix-local"
+  description = "Local Nix repository"
+  notes       = "Internal repository"
+}