+++ examples/resources/artifactory_remote_nix_repository/import.sh
@@ -1,0 +1,1 @@
+terraform import artifactory_remote_nix_repository.my-nix-remote my-nix-remote